refactor(ui): extract revealedSet into useRevealedSet composable (#159)
Deduplicate identical revealedSet computation from WorldMap, MiniMap, and
StatsBar into a shared composable. The composable provides revealedSet,
isRevealed(), and revealedCount from a single reactive source.
